TL-246 [Ndax] Fix subscription objects
The Ndax CCXT Pro handleSubscriptionStatus function searches through `client.subscriptions` to find the id of the subscription that was requested. The `watchOrderBook` subscription populated this dict with subscription objects containing the id. The `authenticate` and `watchAccountEvents` populated it with a future or just `True`. The `handleSubscriptionStatus` function crashes when it calls the `indexBy` function on `client.subscriptions` because `indexBy` doesn't know how to get the `id` from a future or a boolean. Update `authenticate` and `watchAccountEvents` to populate with a structure similar to `watchOrderBook` even though those functions don't really need the structure.